feature{folders(draft-1)}

+++ crud/delete.py
@@ -5,6 +5,7 @@ from threading import Thread
 from werkzeug.exceptions import BadRequest
 from sqlalchemy.exc import IntegrityError
 from werkzeug.security import generate_password_hash,check_password_hash
+import read
 import sys
 sys.path.insert(0, '../models/')
 import os
@@ -23,12 +24,23 @@ def deletefile(fileid,currentuser):
     """
     session=DBSession()
     try:
-        filedetails=session.query(Filedetails).filter_by(fileid=fileid).filter_by(userid=currentuser.userid).one()
-        session.delete(filedetails)  
-        session.commit()
-        os.remove('static/fileuploadfolder/'+fileid)
-        return 'Successfully deleted the file!'
-    except NoResultFound:
-        return 'Failed deleting the file!'
-    except OSError:
-        return 'No File Found!'
+        filedetails = session.query(Filedetails).filter_by(fileid=fileid).filter_by(userid=currentuser.userid).one()
+        files = [filedetails]
+
+        if filedetails.filetype == 'd':
+            files.extend(read.u_getdescendents(fileid))
+        
+        if len(files) > 0:
+            for i in files:
+                filedetails=session.query(Filedetails).filter_by(fileid=i.fileid).filter_by(userid=currentuser.userid).one()
+                session.delete(filedetails)
+            for i in files:
+                if i.filetype == 'f':
+                    os.remove('static/fileuploadfolder/'+i.fileid)
+            session.commit()
+            return 'Successfully deleted'
+        else:
+            return 'Not found'
+    except Exception as err:
+        print err
+        return 'Deletion Failed'
